discordpy-startup/requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
dj-database-url==0.5.0
Django==3.1.1
gunicorn==20.0.4
idna==2.10
Janome==0.4.1
lxml==4.5.2
multidict==4.7.6
numpy==1.19.2
pandas==1.1.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
yarl==1.5.1
psycopg2==2.7.6.1
